Move URI handling into its own file
