+++ extensions/positron-r/package.json
@@ -680,8 +680,5 @@
     },
     "minimumRVersion": "4.2.0",
     "minimumRenvVersion": "1.0.9"
-  },
-  "enabledApiProposals": [
-    "terminalDataWriteEvent"
-  ]
+  }
 }
