+++ extensions/positron-r/tsconfig.json
@@ -24,7 +24,6 @@
 	"include": [
 		"src/**/*",
 		"../../src/positron-dts/positron.d.ts",
-		"../../src/vscode-dts/vscode.d.ts",
-		"../../src/vscode-dts/vscode.proposed.terminalDataWriteEvent.d.ts"
+		"../../src/vscode-dts/vscode.d.ts"
 	]
 }
